Search for closed and departed bookings together
When the turnarounds feature is present in the API, departed bookings will automatically become closed bookings. As we want this to remain invisible to the user until the feature is fully developed, we search for closed and departed bookings together. We accept some loss of atomicity (a booking could transition statuses between calls to the API) as unlikely to cause problems in practice