Add generateConflictBespokeError helper
We add this helper function for creating a bespoke error for conflicts over a bedspace. We do this by inspecting the error message given by the API, and retrieving the ID of the booking/lost bed creating the conflict.

We also add a component for testing its output within integration and E2E tests.