Update booking info component
We update our booking info component to handle bookings with the closed status. For bookings with the departed status, we need to allow them to be displayed as both Departed and 'Departed - Turnaround In Progress'. In our E2E tests, our test runner may have created a booking as departed, but it may be been advanced to closed by the API.

We handle this by reading the status from the page under test, and if the booking is has the status departed but is showing as Departed (rather than 'Departed - Turnaround In Progress'), we update the booking status.