Update BedspaceShowPage
We update shouldShowBookingDetails in BedspaceShowPage to handle bookings with the closed status. For bookings with the departed status, we need to allow them to be displayed as both Departed and 'Departed - Turnaround In Progress'. In our E2E tests, our test runner may have created a booking as departed, but it may be been advanced to closed by the API