Extends booking report controller
Extends controller, view and errors file to accommodate the new month and year fields in the booking report form.

Note that we are validating the form input in the UI - this is because the existing method for piping the response directly from the API makes it difficult for us to access error messages relating to the query params.