Fix ReferenceDataFactory using mocked filterCharacteristics
Our PremisesService and BedspaceService tests were (indirectly) using ReferenceDataFactory, but also mocking characteristicUtils, which is used by ReferenceDataFactory. This meant when ReferenceDataFactory called filterCharacteristics, it was getting undefined, and passing it into faker.helpers.arrayElement, which as of faker 8.0.0 throws an exception on that input.

Our solution is to duplicate filterCharacteristics within ReferenceDataFactory. An alternative would be for our PremisesService and BedspaceService tests to only mock the parts of characteristicUtils they need to, but this assumes they have some knowledge of the inner workings of ReferenceDataFactory (plus that its called at all) to know to need to do this, and is fragile should any other test want to mock characteristicUtils. We accept the repetition of filtering logic as the lesser evil.