Add getApplicationData
Mirroring Approved Premises, we add getApplicationUpdateData and getApplicationSubmissionData for wrangling session application data for transmission to the API. We will use this when we update ApplicationService to be more in line with its Approved Premises counterpart.

Since TA applications do not have the first-class fields that AP applications do, these functions do little for now, but we expect them to expand as the TA Apply feature develops.