Remove test dependency on application release date
In our integration tests, we currently generate mock applications with release date data. As there is no way to enter a release date in the current application form, this does not reflect real data. We update our integration tests to generate applications without a release date, and update ListPage to handle the case where a release date is missing.

We need this change before adding E2E tests for Apply, as applications will only have the data that can be entered into the application form.