Add attach documents task to Apply journey
We move our attach documents task back into the Apply journey, making it part of our new referrals and documents section.

We update our schema and our applicationData.json to include this task, and restore testing of this task in ApplyHelper. Note that we do not check this task when checking testing the check your answers screen within an E2E test. The result is too dependent on data from the API to sensibly test.