Provide section and task data to application pages
We will want to display the task name or the section name within the pages of our Apply form. We first add a helper function, getSectionAndTask(). This is modelled after getPage(), similarly taking a boolean to specify if we are interested in Apply or Assess, and throwing a similar error.

We then use getSectionAndTask() within PagesController to pass section and task data into the rendered template. We must now update some of our existing templates, which expect 'task' to be a string, rather than an object.