Update CheckYourAnswersPage test helper
An Apply page can show its results in the check your answers screen in two forms - a simple set of key/value pairs shown as a GDS summary list. Or, if the value is itself an array of sets of key/value pairs, a GDS summary list containing one or more embedded GDS summary lists. This commit updates shouldShowAnswersForTask() to handle the latter where before it only handled the former