Update getProbationPractitionerName()
We remove an unused optional parameter. This parameter was originally here to keep this function consistent with getArrivalDate(), but these have very different use cases, so divergence is fine. We'll soon be adding more similar functions alongside getProbationPractitionerName() and we don't want to keep this cruft around for all of those