Add completeYesNoInputWithDetailFromPageBody() integration test helper
A common pattern in our Apply form pages is two radio buttons labelled "Yes" and "No”, with a requirement to enter further detail if "Yes" is selected. We have some infrastructure for supporting this case imported from Approved Premises, but nothing for it specifically in our integration tests. We add this helper function for when this pattern arises