Only ask HMPPS Auth for the current user on a new session
Previously we were calling the `GET /auth/api/user/me` endpoint in HMPPS Auth as part of every request.

I think the current need for this is to get the user name for the service header and potentially sending data to HMPPS Audit.

We have seen in production that we have requests that are taking 10-30 seconds and timing out due to closed connections to HMPPS Auth.

Before we escalate this availability issue to them, I think we have an opportunity to hit it once per session rather than once per request. This should improve performance.