Add safeguarding and vulnerability Apply page
We add a safeguarding and vulnerability Apply page, along with a navigation Nunjucks partial, for the tabbed navigation we use on this page and subsequent pages in the same section.

Note that we set columnClasses in our template so that our form fills the width of the page, but then enclose all but the navigation in a two-thirds column. This is so the form elements on this page fill the same width as on other pages, but the navigation tabs are given enough width so as to not wrap onto multiple lines.