Update getSections()
We have a common pattern in our Apply form infrastructure - check if we're dealing with an application or an assessment. In the former case, use Apply.sections(), in the latter case, use getSections() to return the relevant Assess sections.

We can remove some duplication by updating getSection() to take a union type of Assessment and Application, and move the responsibility for determining which type we're detailing with into getSection() itself.

Having made this change, we can simplify our reviewSections() tests, as we can in good conscience rely on getSections() doing the right thing, rather than testing both the Apply and Assess case for this function.

And in our TaskListService tests, we can mock getSections() rather than Apply, which is slightly more straightforward.