Add forPagesInTask()
Both when rendering the check your answers page, and generating the document when submitting a finalised application, we will want to iterate through the form pages in the order a user would journey through them, skipping and pages the user would have skipped.

To prevent duplication, we add forPagesInTask(), which will iterate through form pages for a given task, in user-journey order, calling a given callback for each page.

Note that with some additional complexity, we could adapt forPagesInTask() to be usable in getTaskStatus(), but we accept some repetition for the sake of simplicity.