Update getTaskResponsesAsSummaryListItems()
We update getTaskResponsesAsSummaryListItems() to use forPagesInTask(), so that the check your answers page will only contain answers from pages the user journeyed through in their most recent update to the form.

Note that we no longer check that data for the given task exists on the application. Since function is used for the check your answers page, we should be able to depend on the application being complete. And we do not currently have any tasks that can be skipped. If we ever do want to allow tasks to be skipped, in the same way pages can be skipped in some circumstances, that can be more robustly handled outside this function, ensuring getTaskResponsesAsSummaryListItems() is only ever called for tasks that are part of the user-journey.