Update formPageCheckboxes Nunjucks macro
We update our formPageCheckboxes macro, used for presenting checkboxes in Apply pages, to append square brackets to the field name. Previously, if multiple boxes were checked, these would be received by the page class as an array, but if only one box was checked, this would be received as a string. Appending square brackets to the field name ensures we always receive an array, even if it contains only one element.

This change brings formPageCheckboxes in line with the taCheckboxes macro, used outside of Apply.