Update getApplicationUpdateData() and getApplicationSubmissionData()
We update these methods to return an UpdateApplication/SubmitApplication with the type field specified. As the type field will soon be required, these functions either need to include this field, or we need to change the return type. Opting for the former, we can now remove the addition of this field in ApplicationClient, keeping the responsibility for setting the type ultimately within ApplicationService, which calls getApplicationUpdateData()/getApplicationSubmissionData()