Update BedSearchController
The serviceName field on BedSearchParameters will soon be required, causing the cast to BedSearchParameters within BedSearchController to fail. To support this change, we make it explicit that the search() method on BedspaceSearchService takes a BedSearchParameters with the serviceName omitted, and we remove the failing and potentially misleading cast in BedSearchController. Note that responsibility for adding the serviceName field remains within BedspaceSearchService