Add needs Apply page
This is a page with a series of checkboxes, one for each need, with a field for additional details that must be completed if the checkbox is checked. We also allow the user to select "none of the above”. We use the ‘exclusive' flag on this checkbox, so the other checkboxes are cleared if this is selected. As per GDS guidelines, we also check that this box hasn't been selected alongside other checkboxes, which is possible if Javascript has been disabled in the user's browser