Update dateIsBlank()
We update dateIsBlank() to only examine the date fields with a given prefix. In our release type page, we will want to show distinct errors depending on which date has not been entered, which would not be possible with the previous implementation of dateIsBlank().

We now must update our existing Apply pages that use dateIsBlank(). Note we also change the type of the first parameter to a Partial<>, to save casting within the users of dateIsBlank().