Update date handling in Apply pages
Previously, an Apply page class would receive individual date components as part of its body, and this would indirectly be transformed to an ISO string by a call in the errors() method to dateAndTimeInputsAreValidDates(). This only worked by luck, and relied on dateAndTimeInputsAreValidDates() mutating the object it received, despite it being non-obvious that it performs any mutation.

We now update our page classes to handle date transformation in a body setter, following a pattern established in Approved Premises. Note that body is set within the @Page decorator, so we can rely on the setter being called and the date transformation being applied.