Add adjudications Apply page
We add a page for fetching and displaying adjudication information information. As part of this, we add a navigation partial that will be used by this, and the ACCT alerts page that will follow, for displaying the tab interface for these pages.

Note we add a mapAdjudicationsForPageBody() method, that we use to decouple the API derived type from the body type, that will be used in the Apply schema. We use this to give stronger guarantees about the data we will be storing for this page.