Add ACCT alerts Apply page
We add a page for fetching and displaying ACCT information information. For the tab interface, we use the navigation partial we added in the previous commit.

Note we add a mapAcctAlertsForPageBody() method, that we use to decouple the API derived type from the body type, that will be used in the Apply schema. We use this to give stronger guarantees about the data we will be storing for this page.