Add dateInputHint() helper function
We add this helper function for generating hint text for date input fields, and expose it for use in Nunjucks templates. This will help us generate an example date that is valid for the date input it appears with. For future dates, generating the date dynamically saves us from either having to use dates absurdly far in the future, or needing to update the dates later. We take the approach of using a fixed day and month and only varying the year to avoid inadvertently presenting any inappropriate dates