Invalidate Check Your Answers when changing Apply answers
We force the Apply users to go through the Check Your Answers again if they change any of their answers after having completed Check Your Answers.

First we add a pageBodyShallowEquals() function to allow us to check if the body data of an Apply page has changed. We accept that this will give false positives for pages that contain data beyond flat data submitted from a form (for instance, pages that import data from an external service), but we judge this as acceptable for our current purpose.

We then update ApplicationsService to use pageBodyShallowEquals() to check if the check your answers section should be invalidated.