Update Apply task list page error handling
We update ApplicationsController to make the error handling around submitting an application follow the conventions of the rest of the service, with an error in submit() resulting in a call to catchValidationErrorOrPropogate() and a redirect to the form page. This fixes an issue that caused the task list to disappear if the user attempted to submit the application without checking the confirmation checkbox.

Note in our integration test we take advantage of the change to shouldShowErrorMessagesForFields() we made earlier, and the addition of clickApplication() to ListPage.