Add sentence length Apply page
We add a sentence length Apply page, for entering the sentence length in years, months, weeks, and days. We add validation to ensure at least one of these is entered.

As this page involves a series of text boxes arranged side-by-side, we borrow styling and HTML from the GDS date component. We want errors to be presented for the group of inputs as a whole, requiring us to handle errors 'by hand' in our Nunjucks template.