Move DashboardController and use LandingController
We reorganise DashboardController to live with the manager controllers, and change its path to /dashboard. We then assign LandingController to the root path, so a user reaching the service domain will be redirected to either the dashboard or the applications start page, depending on their role.

We can now integration test the landing controller. We also update some of our existing integration tests to handle the user reaching the application start page directly after login.