Update integration tests to use user roles
Ahead of restricting certain endpoints to certain user roles, we update our integration tests to be run with a mock user of the appropriate role for that test. Though not yet used, we preemptively add setupTestUserWithoutRole() so we can test behaviour when the application is used by someone who is neither an assessor nor a referrer.

We also set our wiremock test user to have both roles, for ease of manual testing.