Add 'show' method to Assessments Controller
This will be used to render the view of the assessment.
I have opted to have the actions for the assessment and the status badge
mapped here, following from the BookingController's precedent but
it may be more appropriate to do this in the view or the even the service.
I add the route in the manage routes file as this is where the other assessment
route lived but I think we may want an 'assess' specific routes file in time