Add 'show' view and test
Here we add the view. In time it will display the content of the assessment
but for now it just displays the person's name, the assessment's status
and the list of actions that can be taken given that assessment's status.

We also test that all the content we expect to be there is