Disable API retries for PUT/POST
Following on with CAS1’s finding that this can create up to 3 duplicate records in the case of API failures[1].

We introduce the COMMUNITY_ACCOMMODATION_API_TIMEOUT_RESPONSE environment variable in the same way as CAS1. Right now they are all configured the same however we may want to refine this in future.

[1] https://github.com/ministryofjustice/hmpps-approved-premises-ui/pull/981