Update forPagesInTask()
We update forPagesInTask() to throw an error if it encounters a page that has no data, or a page that has errors. Our motivation for this is to prevent silent failures when our Apply form changes while a user has part way completed the form. This could otherwise lead to a submission that doesn't match what the user sees on the check your answers page