Add version checking to risk management plan Apply page
The previous version of this page has been deployed and is being used by users. Once this updated version is deployed, we want to ensure that any users that have not yet submitted the form are forced to revisit this page. We do this by adding a hidden "version" field to the Nunjucks template, and checking it in our errors() method.

As this version field will not be present in data generated by the previous version of this page, errors() will return an error until the user revisits this page. And thanks to our changes in pagesForTask(), they will encounter an error if they attempt to view check your answers, or submit the application before doing so. And thanks to our fix in getTaskStatus(), they will not encounter an error in practice, unless they are very unlucky with the timing of a deploy. Instead, the task list will show the task containing this page as incomplete until they revisit it.