Update new booking journey to use query strings
Our create booking journey goes against the usual patterns of our service, as it collects data over two steps before submitting this data to the API. Previously, this data was POSTed, and then stored in a flash message for if we needed to go back a step.

This is unsuitable for when we move to three steps. It makes going back to a middle step difficult, and it is difficult to reason about when flash message data will be cleared. Flash messages are intended to be transient, lasting across a single redirect, and we would need to go against this.

While it has its downsides, including showing an ugly URL to a user, we move to using URL query strings. This gives us greater predictability for how long user data will persist.