Add assessmentRadioItems()
We add this helper function for generating the radio options for our select assessment page. We return options for all provided assessments, sorted to display most recent first, plus an option to select no assessment, using a well-known "no-assessment" string.

Note we make use of the divider option we enabled with our change to the RadioItem type earlier.

Also note that we handle the name not being present, which can occur if the person is a LAO.