Add readyToPlaceForCrn() to AssessmentClient
We add readyToPlaceForCrn() to fetch assessments for our select assessment page. For expediency, we tailor this function to our immediate use-case, rather than adding a more general method than can take a range of statuses. We also update our wiremocks and stubs to support us using this existing API endpoint with query parameters.

We also take this as an opportunity to correct the type of "stubAssessments".