Improve validation on create booking input
We update our selectAssessment() request handler to perform validation on the input received from the new booking page. If a problem is found, we redirect back to the new booking page.

Though not a pattern generally used in this project, we use parameterized tests here, to prevent an unweildy explosion of tests.

We update our E2E and integration tests to reflect errors being caught earlier in our journey.

Update bookingsController.test.ts