Add the phase banner below the users region and display our feedback links
We use the govuk phase banner[1].

We could have added the phaseBanner directly into the header.njk but given we’ve already split out the probationRegionHeader (which is a similar part of the header) out into it’s own file, I followed the same approach.

We put the feedback below the region in line with what looks better and what CAS1 do as a point of consistency.

[1]https://design-system.service.gov.uk/components/phase-banner/