Viewing an individual booking works for an LAO
If the current user doesn't have access to the CRN as they are a limited
access offender, the code before this would error where the dateOfBirth
is undefined and we try to format it.

This change is a simple guard against a person without a dateofbirth so
that the page continues to render.

We also take this opportunity to improve the way a missing person's name
is handled. Right now it'll render and say "Name: ", we improve that.