Listing bookings on a bedspace works when one belongs to a limited access offender
If the logged in user doesn’t have access to view this bookings' person
details, the API will return a limited result set. Instead of erroring
when no date of birth is returned we guard against it being blank.

We also guard against the person's name being blank. Instead of showing
an empty <h3> we add an LAO placeholder.