Add shortAddress() helper function
Constructing a short, one-line address for a premises is a common pattern in our app, so we add a utility function to do this. As per the latest bedspace search designs, we include the premises town if present. Though changing every instance of wanting a short address to use this function is out-of-scope for this PR, in future we should use this function for consistency.

Additionally, we expose PremisesUtils, containing this function, to our Nunjucks templates.