Update bedspace search results page
We update this page to display more detailed results, including the characteristics of each bedspace and its premises. We use the helper functions we added in the previous two commits.

Though not included in the design, we keep the bed count out of caution, until we are certain the the "shared property" attribute is correctly set for each premises.