Prevent person name being stored with submitted application
Given the existence of LAOs, and given that a referrer may have different access rights than the assessor, we should err on the side of caution around embedding personal identifying information in the application itself. To that end, we review and modify where necessary the response() method of each of our application pages to not include the name of the person being referred. We instead use "this person", as already used on our accommodation application sharing page