Set flash messages on res.locals
We update app.ts so that flash messages are set on res.locals rather than res.app.locals. This ensures their scope is a single request, rather than the application as a whole. This fixes a likely culprit for our E2E tests sometimes failing due to flash messages not appearing