Add place context header macro
We add a Nunjucks macro for displaying place context on pages in the user journey from an assessment through to completing a booking. This will display essential information about the person being placed.

We also add PlaceContextHeaderComponent, and add shouldShowplaceContextHeader() to our base Page class to assist integration testing this component.