Add placeUtils
We add a first iteration of placeUtils, with two helper functions, createPlaceContext() and addPlaceContext(). The former is a simple transformation from an Assessment type to a PlaceContext type, so that the content of a place context is abstracted from its users. The latter takes a link and appends it with the ID of the assessment, which we will use for tracking a place context across our user journey.

We also add a factory for our new PlaceContext type.