Add preservePlaceContext() to placeUtils
We add this function for retrieving place context information from the assessment service, and exposing it to Nunjucks, via res.locals. We want to call this in the request handler of any page we want to be aware of place context. It now makes sense to expose addPlaceContext() to Nunjucks, with the placeContext parameter being populated automatically.

We swallow any errors from the assessment service, as place context is never essential to a page, and errors here should not prevent the user from progressing.

We use declaration merging to add placeContext to the type of res.locals.