Make bedspace search page place context-aware
We update the bedspace search page to display a place context header, and to appended the place context assessment ID to outgoing links to pages that will in future also be place context-aware.

Note that our previous check for whether the user was viewing this page pre- or post-form submission no longer holds, given the possible presence of a place context assessment ID. Therefore, we update it to check for a known search parameter instead.