Add clearPlaceContext()
We are now at the new booking stage of our place journey. If at this stage, the user enters a CRN that differs from the assessment CRN, or selects a different assessment, we take that as a sign that we should clear any place context. We add this method to enable that