Make booking confirm page place context-aware
We update the booking confirm page to display a place context header, and additionally to clear the place context if the assessment ID we receive from the previous page differs from that in the place context.

Note that we do not need to call addPlaceContext() on our backLink value, due to the existing use of appendQueryString().