Render back button outside of main content
It was flagged that too much padding appeared on top of the back
buttons. This is due to padding coming from the `govuk-main-wrapper`
class on the `main` tag, which is the parent of the back link anchor.

To remedy this, this change renders the back link outside of the `main`
tag as the breadcrumbs are technically for navigation and therefore
doesn't need to be included in `main`. This is consistent with other gov
uk pages[1].

[1]
https://www.gov.uk/sign-in-universal-credit