Add helper function timelineItems()
This helper function transforms assessment notes to a format accepted by the MoJ Timeline component, sorted for presentation in reverse-chronological order.

As appears to be standard MoJ practice for user names[1], we use convertToTitleCase() to format the name for display.

Though the API currently only supports user notes, we expect soon it will return non-user notes, so we ensure we do not depend on the presence of createdByUserName on our note object.

[1] https://github.com/ministryofjustice/hmpps-template-typescript/blob/fc402c224eee3eeec0f1f56e1b92e452648be1f7/server/services/userService.ts#L14