Rename assessment 'show' page to 'full' page
As part of the work to introduce a new assessment summary page, this
change renames the current assessment page and its dependencies from
'show' to 'full'.

At present, this mirrors the behaviour of the old page, but once the new
summary page is implemented, we'll update the full page (for example,
the action buttons will be on the new summary page and will be removed
from the full referral page.