Introduce referral summary page
This change adds the referall summary page which displays the POP
details and a link to the full referral. We will add
more information to this page in the future.

Where we were linking to the full referall, we now link to the summary:

- In the referalls table
- Redirecting after updating the referral

In the interest of keeping this commit as small as possible, updates to
the full referral page will be made in following commits.