Update content on booking form page for when no associated referrals exist
This commit changes the content of the page shown when attempting to
make a booking for which a corresponding referral doesn't exist for the
person.

We received a support ticket for the page as the user thought they
had encountered an error.

This new content should help give the user better guidance on what to do
next.