Update bedspace search E2E test
We update our bedspace search E2E test to ensure that the search results we pass into BedspaceSearchPage contain zero overlaps. Since our test creates a new premises and bedspace to search for, we can guarantee this will match what the test sees on the page.

This fixes an issue where our tests would only pass on the one-in-six chance that the result generated by bedSearchResultFactory contained zero overlaps.