Hide system notes from assessments
An upcoming API PR[1] will add system notes to assessments. To prevent these showing in a way that will be confusing to users, we hide them until we can implement display of system notes properly.

Though the upcoming PR will send a 'type' field with each note, we cannot use that yet, and to resolve this chicken-and-egg situation, we check if 'message' is populated to determine if this is a system note.

[1] https://github.com/ministryofjustice/hmpps-approved-premises-api/pull/919