Add gender identity to E2E test data
We update our data that mirrors the person used in E2E tests to include the expected gender identity of the person. We will require this data to be correct, as it is used as part of testing the place context header.

Note that the use of "Jedi" as gender identity for our local test user is driven by the need to match the default person from the Community API[1].

[1] https://github.com/search?q=repo%3Aministryofjustice%2Fcommunity-api%20jedi&type=code