E2E test place feature
We now expand our existing E2E test to cover placing a generated assessment. We first open the assessment generated from the application created in the previous scenario. We then mark the assessment as ready to place. And then reuse the place helper we added for integration testing this feature to book a bedspace for this assessment