Update BookingSelectAssessmentPage
We update how this test helper selects a radio button for an assessment. We now use the label text, rather than the assessment ID, which we will not know in our E2E tests. We also make it the responsibility of the lower level methods in Page to account for the extra spaces GDS wraps around radio button label text