Update messages macro
We update our messages macro to handle success messages that can either be plain text or HTML, and ensure text is escaped correctly. This follows a similar pattern to many GDS components, where a 'text' or an 'html' field may be specified.

For completeness, we make similar changes to info messages to match. We also add a new type, to encode the data shape this macro expects.