Conditionally render the ACCT alert expiry date when rendering the form responses
We received two support tickets[1] from users who were unable to submit
their application. We can see issues[2] coming through to sentry on the
'Check your answers' page before submitting a referral.

Upon inspection, it is possiblle for an ACCT alert to have a missing
`dateExpires` property, so this change adds a null guard against that
scenario. We render an empty string, which aligns with how we handle it
elsewhere[3].

[1]
https://mojdt.slack.com/archives/C04NJ6BATUL/p1693409827077599?thread_ts=1693409758.827739&cid=C04NJ6BATUL
https://mojdt.slack.com/archives/C04NJ6BATUL/p1693411279302619?thread_ts=1693409758.827739&cid=C04NJ6BATUL

[2]
https://ministryofjustice.sentry.io/issues/4347240074/events/39b224f1720e4e10af3e3aa7d4e6d8dc/

[3]
https://github.com/ministryofjustice/hmpps-temporary-accommodation-ui/blob/main/server/views/applications/pages/assess-placement-risks-and-needs/prison-information/acct-alerts.njk#L27