Conditionally show archived info message when viewing a property
When updating or creating a property with the `archived` status, we
render two notification banners on success. This fails accessibility
tests as they have duplicate ARIA label ids.

To remedy this, we only render the info banner when no success messages
are present. We also change the success message on update and create to
notify the user that the property is archived.

This is also in line with guidance from the design system, where it's
discouraged to render 2 notification banners.

https://dequeuniversity.com/rules/axe/4.7/duplicate-id-aria
https://design-system.service.gov.uk/components/notification-banner/